src/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-API==2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pymongo==3.10.1
Werkzeug==1.0.1
